pyproject.toml
[project]
name = "opensource-repo-analyser"
version = "0.1.0"
description = "Multi-agent GitHub repository analysis system built with Google ADK, FastAPI, and async Python services."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=0.1.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
    "src",
]
